added -gpu cmdline option; debugging the gpu part in progress

+++ 00_include/struct_binheap.h
@@ -6,7 +6,7 @@ extern "C" {
 #endif
 
 typedef struct {
-    float rmsd;
+    double rmsd;
     int triple_x[3];
     int triple_y[3];
     double quat[4];
